requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# _libgcc_mutex=0.1=main
blas=1.0=mkl
blosc=1.20.1=he1b5a44_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.1.19=h06a4308_0
cairo=1.14.12=h8948797_3
certifi=2020.12.5=py36h5fab9bb_1
cffi=1.14.4=py36h261ae71_0
cloudpickle=1.6.0=py_0
cudatoolkit=9.2=0
cycler=0.10.0=py36_0
cytoolz=0.11.0=py36h7b6447c_0
dask-core=2021.1.1=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
deepdish=0.3.6=pyh9f0ad1d_0
expat=2.2.10=he6710b0_2
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.10.4=h5ab3b9f_0
glib=2.66.1=h92f7085_0
graphite2=1.3.14=h23475e2_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.8.0=py36h989c5e5_3
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
imageio=2.9.0=py_0
intel-openmp=2020.2=254
jasper=2.0.14=h07fcdf6_1
joblib=1.0.0=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.0=py36h2531618_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.3=h2531618_0
lzo=2.10=h516909a_1000
matplotlib=3.1.0=py36h5429711_0
matplotlib-base=3.3.2=py36h817c723_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.2.0=py36h23d657b_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.2=he6710b0_1
networkx=2.5=py_0
ninja=1.10.2=py36hff7bd54_0
numexpr=2.7.1=py36h7c3b610_3
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=py36_0
opencv=3.4.2=py36h6fd60c2_1
openssl=1.1.1i=h27cfd23_0
pcre=8.44=he6710b0_0
pillow=8.1.0=py36he98fc37_0
pip=20.3.3=py36h06a4308_0
pixman=0.40.0=h7b6447c_0
py-opencv=3.4.2=py36hb342d67_1
pycparser=2.20=py_2
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py36h05f1152_2
pytables=3.4.4=py36h4f72b40_1
python=3.6.12=hcff3b4d_2
python-dateutil=2.8.1=py_0
python_abi=3.6=1_cp36m
pytorch=1.2.0=py3.6_cuda9.2.148_cudnn7.6.2_0
pytz=2020.5=pyhd3eb1b0_0
pywavelets=1.1.1=py36h7b6447c_2
pyyaml=5.4.1=py36h27cfd23_1
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
scikit-image=0.17.2=py36hdf5156a_0
scikit-learn=0.23.2=py36h0573a6f_0
scipy=1.5.2=py36h0b6359f_0
setuptools=52.0.0=py36h06a4308_0
sip=4.19.8=py36hf484d3e_0
six=1.15.0=py36h06a4308_0
sqlite=3.33.0=h62c20be_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2020.10.1=py36hdd07704_2
tk=8.6.10=hbc83047_0
toolz=0.11.1=pyhd3eb1b0_0
torchvision=0.4.0=py36_cu92
tornado=6.1=py36h27cfd23_0
tqdm=4.56.0=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
